made profiels directory and moved files from auth
